Migrate keysToSnakeCase to utils
